`\gtrifdateorplacedefined` macro
Evaluates to true if either date or place is defined. This allows `\gtrPrintEvent@prefixdateplace` to print events that contain only a place, but no date. Can for example be used if a person has no known birth date, but the address is of interest in conjunction with a baptismal date; likewise for death place + burial date:

```
{
	female,
	name = {Anne Sophie \surn{Tofte}},
	birth = {}{Nørrevold},
	baptism = {1790-06-28}{Trinitatis kirke},
	death = {1871-05-31}{Rigensgade 8},
	burial = {1871-06-04}{Skt. Paul kirke},
}
```

Renders as (symbols approximated):

```
Anne Sophie Tofte
* i Nørrevold
~ 28. juni 1790 i Trinitatis kirke
† 31. maj 1871 i Rigensgade 8
[] 4. juni 1871 i Skt. Paul kirke
```